labki10.04

+++ fizyka.cpp
@@ -37,7 +37,7 @@ void CFizyka::Odbicie(float alfa_n) //odbicie od sciany charakteryzowanej za pom
 void CFizyka::Aktualizuj(int czas_aktualny) //zmienia polozenie obiektu na podstawie aktualnego czasu
 {
 
-	float delta_t = czas_aktualny - czas, v_x, v_y;
+	double delta_t = czas_aktualny - czas, v_x, v_y;
 	if (delta_t>1000) delta_t = 100;//dla przerwy dluzszej niz 1s nie przeprowadzana jest aktualizacja
 	v_x = v*cos(alfa_v / 180.0*PI);
 	v_y = v*sin(alfa_v / 180.0*PI);
